requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
exceptiongroup==1.3.0
    # via anyio
grpcio==1.75.1
    # via qdrant-client
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via qdrant-client
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
numpy==2.2.6
    # via qdrant-client
portalocker==3.2.0
    # via qdrant-client
protobuf==6.32.1
    # via qdrant-client
pydantic==2.11.9
    # via qdrant-client
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.1
    # via -r requirements.in
qdrant-client==1.15.1
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via qdrant-client
